requirements.txt
Django==2.0.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pep8==1.7.1
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
